refactor: move shared code out of tests
