+++ test/nuts/shared/isNameObsolete.ts
@@ -1,0 +1,20 @@
+/*
+ * Copyright (c) 2020, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+import { AuthInfo, Connection } from '@salesforce/core';
+
+export const isNameObsolete = async (username: string, memberType: string, memberName: string): Promise<boolean> => {
+  const connection = await Connection.create({
+    authInfo: await AuthInfo.create({ username }),
+  });
+
+  const res = await connection.singleRecordQuery<{ IsNameObsolete: boolean }>(
+    `SELECT IsNameObsolete FROM SourceMember WHERE MemberType='${memberType}' AND MemberName='${memberName}'`,
+    { tooling: true }
+  );
+
+  return res.IsNameObsolete;
+};
